Initial commit of more flexible auto-grade policies.
This implementation allows assigning point values, either all the same, or on an assignment-by-assignment basis, for auto grading, at the course and section levels. A tool is provided
to help with this, along with helping with the date-overrides for section assignments.
Part/question level points will be dealt with as-when.

Note that the on-disk filename has changed, check the updated documentation.


git-svn-id: https://repos.nextthought.com/svn/nti-svn/nti.app.assessment/trunk@45156 3dfb626b-079f-4ab1-a9f6-4654e5377508